# packages/client-web/src/generated/models.ts
// <auto-generated>
// Generated by tools/generate-clients.mjs from packages/contracts/openapi/settleora.v1.yaml.
// Do not edit this file by hand.
// </auto-generated>

/**
 * Bounded setup status for first-owner local bootstrap.
 */
export interface BootstrapStatusResponse {
  /**
   * True only when no auth account exists.
   */
  bootstrapRequired: boolean;
}

/**
 * First-owner local bootstrap request for a fresh deployment. This is not general public registration.
 */
export interface BootstrapLocalOwnerRequest {
  /**
   * Submitted local account identifier, normalized server-side consistently with local sign-in.
   */
  identifier: string;
  /**
   * Submitted local account password. The API never stores this plaintext value.
   */
  password: string;
  /**
   * Initial user profile display name after server-side trimming.
   */
  displayName: string;
  /**
   * Optional initial default currency. Explicit null is accepted.
   */
  defaultCurrency?: CurrencyCode | null;
}

/**
 * Safe first-owner bootstrap response. It excludes session tokens, password material, credential metadata, provider payloads, audit metadata, storage paths, and unrelated records.
 */
export interface BootstrapLocalOwnerResponse {
  userProfile: SelfUserProfileResponse;
  roles: ("owner" | "admin" | "user")[];
}

/**
 * Local-account sign-in request. Source, network, and policy bucket details are derived server-side.
 */
export interface LocalSignInRequest {
  /**
   * Submitted local account identifier.
   */
  identifier: string;
  /**
   * Submitted local account password. The API never stores this plaintext value.
   */
  password: string;
  /**
   * Optional display label for the newly created session.
   */
  deviceLabel?: string;
}

/**
 * Minimal refresh-capable local sign-in success response. Raw credential material is returned only once.
 */
export interface LocalSignInResponse {
  session: RefreshSessionAccessSession;
  refreshCredential: RefreshSessionCredential;
}

/**
 * Public refresh request. Credential lookup, account, session-family, status, expiry, and revocation policy are resolved server-side.
 */
export interface RefreshSessionRequest {
  /**
   * Submitted raw refresh-like credential. The API never stores this raw value.
   */
  refreshCredential: string;
  /**
   * Optional display label for the replacement access session.
   */
  deviceLabel?: string;
}

/**
 * Minimal refresh success response. Raw credential material is returned only once.
 */
export interface RefreshSessionResponse {
  session: RefreshSessionAccessSession;
  refreshCredential: RefreshSessionCredential;
}

/**
 * Replacement opaque access session returned from refresh.
 */
export interface RefreshSessionAccessSession {
  id: string;
  /**
   * Opaque raw access-session token returned once on refresh success.
   */
  token: string;
  expiresAtUtc: string;
}

/**
 * Replacement refresh-like credential returned from refresh.
 */
export interface RefreshSessionCredential {
  /**
   * Raw refresh-like credential returned once on refresh success.
   */
  token: string;
  idleExpiresAtUtc: string;
  absoluteExpiresAtUtc: string;
}

/**
 * Minimal current authenticated actor, profile, session, and system-role summary.
 */
export interface CurrentUserResponse {
  authAccountId: string;
  userProfile: CurrentUserProfile;
  session: CurrentUserSession;
  roles: ("owner" | "admin" | "user")[];
}

/**
 * Linked active user profile summary for the authenticated account.
 */
export interface CurrentUserProfile {
  id: string;
  displayName: string;
  defaultCurrency: CurrencyCode | null;
}

/**
 * Validated current session summary.
 */
export interface CurrentUserSession {
  id: string;
  expiresAtUtc: string;
}

/**
 * Safe self-profile response for the authenticated actor.
 */
export interface SelfUserProfileResponse {
  id: string;
  displayName: string;
  defaultCurrency: CurrencyCode | null;
  createdAtUtc: string;
  updatedAtUtc: string;
}

/**
 * Safe self-profile update. Profile identity is derived from the authenticated actor, not from client input.
 */
export interface UpdateSelfUserProfileRequest {
  /**
   * Display name after server-side trimming. Blank or whitespace-only values are rejected.
   */
  displayName?: string;
  /**
   * Optional default currency. Explicit null clears the value.
   */
  defaultCurrency?: CurrencyCode | null;
}

/**
 * Safe active session metadata for the authenticated account. The list is capped by the API.
 */
export interface SessionListResponse {
  sessions: SessionSummary[];
}

/**
 * Bounded session metadata that excludes token material, hashes, network data, user-agent data, and revocation details.
 */
export interface SessionSummary {
  id: string;
  isCurrent: boolean;
  status: "active";
  issuedAtUtc: string;
  expiresAtUtc: string;
  lastSeenAtUtc: string | null;
  deviceLabel: string | null;
}

/**
 * ISO-4217 style currency code.
 */
export type CurrencyCode = string;

/**
 * Decimal-safe money value with an attached currency.
 */
export interface Money {
  /**
   * Decimal-safe amount represented as a string to avoid floating point loss.
   */
  amount: string;
  currency: CurrencyCode;
}

/**
 * Placeholder rounding modes for future policy-driven money handling.
 */
export type RoundingMode = "up" | "down" | "nearest" | "bankers";

/**
 * Placeholder sync state values shared by local and server-mode clients.
 */
export type SyncState = "queued" | "synced" | "conflict" | "failed";

/**
 * Stable storage object reference that does not expose physical filesystem paths.
 */
export interface StorageObjectRef {
  /**
   * Opaque storage object identifier.
   */
  id: string;
  contentType?: string;
  sizeBytes?: number;
}

/**
 * RFC 7807 / RFC 9457 style problem response.
 */
export interface ProblemDetails {
  type?: string;
  title?: string;
  status?: number;
  detail?: string;
  instance?: string;
}

/**
 * Placeholder OCR job status values for future worker contract design.
 */
export type OcrJobStatus = "queued" | "processing" | "completed" | "failed";

export interface GetHealthResponse {
  status: string;
}

export interface GetHealthReadyResponse {
  status: "ready";
  service: string;
  checks: { postgres: "ok"; rabbitmq: "ok"; storage: "ok" };
}