# packages/contracts/openapi/settleora.v1.yaml
openapi: 3.1.0
info:
  title: Settleora API
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    OpenAPI contract for Settleora scaffold materialization. Generated mobile
    and web clients are produced from this file, and generated output
    must not be manually edited. This contract currently covers scaffold health
    checks, first-owner local bootstrap status/creation for fresh deployments,
    refresh-capable local sign-in, public refresh, current-session sign-out,
    current-account sign-out-all, current-account session list and per-session
    revocation, the minimal current-user auth read boundary, and guarded
    self-profile read/update endpoints; it is not the feature contract for
    general public registration, auth middleware, session management beyond
    current-session sign-out, current-account sign-out-all/list/revocation, and
    public refresh, general user management, group endpoints, expenses,
    settlement, sync, OCR upload, or database behavior.
servers:
  - url: /
paths:
...
  /api/v1/auth/bootstrap/status:
    get:
      summary: Bootstrap status
      description: Returns only whether first-owner local bootstrap is required because no auth account exists.
      operationId: getAuthBootstrapStatus
      security: []
      responses:
        "200":
          description: Bounded bootstrap setup status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapStatusResponse"
        "400":
          description: Problem response.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        default:
          description: Problem response.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/auth/bootstrap/local-owner:
    post:
      summary: Bootstrap first local owner
      description: Creates the first local owner account only while no auth account exists. This is not general public registration; clients must sign in normally after bootstrap succeeds.
      operationId: bootstrapLocalOwner
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapLocalOwnerRequest"
      responses:
        "200":
          description: Local owner bootstrap succeeded without returning credential or token material.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapLocalOwnerResponse"
        "400":
          description: The submitted bootstrap request is invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Bootstrap is unavailable because an auth account already exists.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Bootstrap could not be completed.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        default:
          description: Problem response.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  securitySchemes:
    SessionBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  schemas:
    BootstrapStatusResponse:
      type: object
      description: Bounded setup status for first-owner local bootstrap.
      required:
        - bootstrapRequired
      properties:
        bootstrapRequired:
          type: boolean
          description: True only when no auth account exists.
    BootstrapLocalOwnerRequest:
      type: object
      description: First-owner local bootstrap request for a fresh deployment. This is not general public registration.
      additionalProperties: false
      required:
        - identifier
        - password
        - displayName
      properties:
        identifier:
          type: string
          minLength: 1
          maxLength: 320
          description: Submitted local account identifier, normalized server-side consistently with local sign-in.
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 4096
          description: Submitted local account password. The API never stores this plaintext value.
        displayName:
          type: string
          minLength: 1
          maxLength: 160
          description: Initial user profile display name after server-side trimming.
        defaultCurrency:
          oneOf:
            - $ref: "#/components/schemas/CurrencyCode"
            - type: "null"
          description: Optional initial default currency. Explicit null is accepted.
    BootstrapLocalOwnerResponse:
      type: object
      description: Safe first-owner bootstrap response. It excludes session tokens, password material, credential metadata, provider payloads, audit metadata, storage paths, and unrelated records.
      required:
        - userProfile
        - roles
      properties:
        userProfile:
          $ref: "#/components/schemas/SelfUserProfileResponse"
        roles:
          type: array
          items:
            type: string
            enum:
              - owner
              - admin
              - user
...
    SelfUserProfileResponse:
      type: object
      description: Safe self-profile response for the authenticated actor.
      required:
        - id
        - displayName
        - defaultCurrency
        - createdAtUtc
        - updatedAtUtc
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          minLength: 1
          maxLength: 160
        defaultCurrency:
          oneOf:
            - $ref: "#/components/schemas/CurrencyCode"
            - type: "null"
        createdAtUtc:
          type: string
          format: date-time
        updatedAtUtc:
          type: string
          format: date-time
...
    CurrencyCode:
      type: string
      pattern: "^[A-Z]{3}$"
      description: ISO-4217 style currency code.
      examples:
        - USD
...
    ProblemDetails:
      type: object
      description: RFC 7807 / RFC 9457 style problem response.
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
...